Create Rgeoprofile data object

Simple function that ensures that input data is in the correct format required by Rgeoprofile. Takes longitude and latitude as input vectors and returns these same values in list format.

geoData(longitude = NULL, latitude = NULL)

Arguments

longitude
the locations of the observed data in degrees longitude.

latitude
the locations of the observed data in degrees latitude.

Examples

# John Snow cholera data
geoData(Cholera$longitude, Cholera$latitude)

#> $longitude
#>   [1] -0.140139 -0.140066 -0.140037 -0.139700 -0.139693 -0.139605 -0.139576
#>   [8] -0.139539 -0.139473 -0.139400 -0.139349 -0.139334 -0.139334 -0.139312
#>  [15] -0.139312 -0.139224 -0.139210 -0.139180 -0.139107 -0.139034 -0.139034
#>  [22] -0.138961 -0.138902 -0.138858 -0.138837 -0.138815 -0.138793 -0.138756
#>  [29] -0.138734 -0.138690 -0.138646 -0.138639 -0.138551 -0.138507 -0.138441
#>  [36] -0.138427 -0.138397 -0.138361 -0.138354 -0.138346 -0.138295 -0.138288
#>  [43] -0.138280 -0.138273 -0.138251 -0.138236 -0.138229 -0.138222 -0.138200
#>  [50] -0.138163 -0.138141 -0.138141 -0.138105 -0.138105 -0.138068 -0.138061
#>  [57] -0.138002 -0.137988 -0.137973 -0.137936 -0.137929 -0.137900 -0.137885
#>  [64] -0.137885 -0.137849 -0.137849 -0.137841 -0.137841 -0.137819 -0.137812
#>  [71] -0.137768 -0.137717 -0.137709 -0.137688 -0.137666 -0.137614 -0.137607
#>  [78] -0.137556 -0.137548 -0.137527 -0.137497 -0.137468 -0.137461 -0.137431
#>  [85] -0.137424 -0.137395 -0.137395 -0.137380 -0.137380 -0.137373 -0.137344
#>  [92] -0.137344 -0.137314 -0.137248 -0.137219 -0.137219 -0.137117 -0.137095
#>  [99] -0.137095 -0.137065 -0.137014 -0.136992 -0.136985 -0.136963 -0.136919
#> [106] -0.136919 -0.136875 -0.136861 -0.136853 -0.136802 -0.136787 -0.136780
#> [113] -0.136780 -0.136765 -0.136722 -0.136714 -0.136714 -0.136685 -0.136678
#> [120] -0.136656 -0.136656 -0.136634 -0.136612 -0.136604 -0.136582 -0.136575
#> [127] -0.136575 -0.136539 -0.136502 -0.136502 -0.136465 -0.136458 -0.136443
#> [134] -0.136429 -0.136429 -0.136421 -0.136392 -0.136370 -0.136363 -0.136363
#> [141] -0.136363 -0.136348 -0.136319 -0.136312 -0.136304 -0.136282 -0.136260
#> [148] -0.136231 -0.136224 -0.136217 -0.136202 -0.136202 -0.136202 -0.136180
#> [155] -0.136173 -0.136151 -0.136136 -0.136136 -0.136136 -0.136136 -0.136121
#> [162] -0.136099 -0.136099 -0.136085 -0.136085 -0.136063 -0.136063 -0.136056
#> [169] -0.136048 -0.136041 -0.136019 -0.136012 -0.136004 -0.136004 -0.136004
#> [176] -0.135997 -0.135982 -0.135968 -0.135953 -0.135924 -0.135909 -0.135902
#> [183] -0.135895 -0.135887 -0.135865 -0.135865 -0.135858 -0.135851 -0.135836
#> [190] -0.135821 -0.135814 -0.135799 -0.135785 -0.135777 -0.135777 -0.135763
#> [197] -0.135748 -0.135719 -0.135719 -0.135675 -0.135668 -0.135668 -0.135646
#> [204] -0.135638 -0.135587 -0.135573 -0.135573 -0.135565 -0.135536 -0.135536
#> [211] -0.135514 -0.135499 -0.135485 -0.135455 -0.135455 -0.135426 -0.135419
#> [218] -0.135412 -0.135404 -0.135404 -0.135353 -0.135346 -0.135346 -0.135324
#> [225] -0.135251 -0.135251 -0.135236 -0.135207 -0.135199 -0.135192 -0.135177
#> [232] -0.135170 -0.135141 -0.135133 -0.135097 -0.135068 -0.135024 -0.135009
#> [239] -0.135002 -0.134987 -0.134987 -0.134987 -0.134958 -0.134921 -0.134914
#> [246] -0.134914 -0.134892 -0.134863 -0.134833 -0.134826 -0.134804 -0.134797
#> [253] -0.134760 -0.134760 -0.134724 -0.134716 -0.134687 -0.134665 -0.134665
#> [260] -0.134658 -0.134628 -0.134614 -0.134614 -0.134607 -0.134599 -0.134570
#> [267] -0.134548 -0.134541 -0.134533 -0.134526 -0.134519 -0.134489 -0.134482
#> [274] -0.134475 -0.134453 -0.134431 -0.134394 -0.134380 -0.134365 -0.134314
#> [281] -0.134306 -0.134306 -0.134284 -0.134255 -0.134248 -0.134211 -0.134189
#> [288] -0.134182 -0.134182 -0.134182 -0.134160 -0.134145 -0.134123 -0.134087
#> [295] -0.134072 -0.134065 -0.134028 -0.134014 -0.133999 -0.133933 -0.133919
#> [302] -0.133904 -0.133845 -0.133845 -0.133831 -0.133794 -0.133787 -0.133743
#> [309] -0.133721 -0.133662 -0.133575 -0.133516 -0.133472 -0.133436 -0.133370
#> [316] -0.133318 -0.133297 -0.132960 -0.132777 -0.132770 -0.132528
#> 
#> $latitude
#>   [1] 51.51111 51.51324 51.51120 51.51348 51.51293 51.51352 51.51498 51.51068
#>   [9] 51.51306 51.51404 51.51300 51.51288 51.51344 51.51279 51.51361 51.51352
#>  [17] 51.51294 51.51285 51.51322 51.51292 51.51334 51.51296 51.51362 51.51333
#>  [25] 51.51369 51.51353 51.51364 51.51337 51.51358 51.51309 51.51317 51.51278
#>  [33] 51.51339 51.51219 51.51315 51.51274 51.51310 51.51322 51.51304 51.51332
#>  [41] 51.51386 51.51365 51.51326 51.51296 51.51359 51.51336 51.51240 51.51291
#>  [49] 51.51329 51.51188 51.51343 51.51365 51.51229 51.51242 51.51326 51.51238
#>  [57] 51.51232 51.51344 51.51404 51.51311 51.51335 51.51471 51.51305 51.51330
#>  [65] 51.51325 51.51562 51.51228 51.51300 51.51175 51.51320 51.51314 51.51285
#>  [73] 51.51475 51.51234 51.51226 51.51295 51.51238 51.51265 51.51312 51.51281
#>  [81] 51.51223 51.51300 51.51370 51.51356 51.51267 51.51227 51.51302 51.51263
#>  [89] 51.51350 51.51372 51.51221 51.51258 51.51305 51.51331 51.51255 51.51492
#>  [97] 51.51448 51.51259 51.51311 51.51467 51.51243 51.51315 51.51262 51.51331
#> [105] 51.51422 51.51445 51.51318 51.51334 51.51268 51.51480 51.51320 51.51400
#> [113] 51.51424 51.51337 51.51390 51.51272 51.51322 51.51409 51.51340 51.51260
#> [121] 51.51380 51.51275 51.51323 51.51255 51.51487 51.51201 51.51343 51.51278
#> [129] 51.51308 51.51381 51.51237 51.51321 51.51281 51.51233 51.51259 51.51453
#> [137] 51.51328 51.51275 51.51251 51.51290 51.51537 51.51223 51.51428 51.51331
#> [145] 51.51302 51.51457 51.51297 51.51431 51.51333 51.51234 51.51272 51.51354
#> [153] 51.51449 51.51229 51.51387 51.51268 51.51223 51.51282 51.51336 51.51356
#> [161] 51.51398 51.51389 51.51451 51.51276 51.51293 51.51215 51.51358 51.51260
#> [169] 51.51339 51.51477 51.51290 51.51255 51.51187 51.51295 51.51454 51.51393
#> [177] 51.51361 51.51403 51.51293 51.51298 51.51245 51.51363 51.51196 51.51406
#> [185] 51.51295 51.51320 51.51397 51.51192 51.51300 51.51315 51.51366 51.51408
#> [193] 51.51311 51.51186 51.51246 51.51503 51.51306 51.51298 51.51434 51.51293
#> [201] 51.51371 51.51427 51.51445 51.51288 51.51373 51.51416 51.51437 51.51468
#> [209] 51.51276 51.51431 51.51208 51.51376 51.51405 51.51203 51.51269 51.51213
#> [217] 51.51416 51.51360 51.51394 51.51511 51.51265 51.51342 51.51408 51.51362
#> [225] 51.51340 51.51516 51.51366 51.51391 51.51320 51.51192 51.51347 51.51175
#> [233] 51.51315 51.51272 51.51309 51.51352 51.51247 51.51239 51.51383 51.51201
#> [241] 51.51209 51.51544 51.51278 51.51336 51.51242 51.51353 51.51398 51.51441
#> [249] 51.51245 51.51233 51.51479 51.51376 51.51257 51.51336 51.51331 51.51363
#> [257] 51.51441 51.51242 51.51417 51.51322 51.51172 51.51240 51.51317 51.51397
#> [265] 51.51197 51.51207 51.51242 51.51108 51.51309 51.51231 51.51343 51.51413
#> [273] 51.51582 51.51303 51.51408 51.51298 51.51293 51.51426 51.51400 51.51284
#> [281] 51.51047 51.51374 51.51100 51.51444 51.51365 51.51382 51.51318 51.51217
#> [289] 51.51356 51.51432 51.51429 51.51369 51.51346 51.51435 51.51432 51.51356
#> [297] 51.51197 51.51347 51.51192 51.51237 51.51444 51.51435 51.51283 51.51440
#> [305] 51.51446 51.51235 51.51243 51.51449 51.51443 51.51451 51.51454 51.51307
#> [313] 51.51452 51.51162 51.51454 51.51308 51.51316 51.51317 51.51106 51.51297
#> [321] 51.51097
#> 

# simulated data
sim <-rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude =
51.5235505, alpha=1, sigma=1, tau=3)
geoData(sim$longitude, sim $latitude)

#> $longitude
#>  [1] -0.055608599 -0.084033810 -0.052196790 -0.058607065 -0.050691397
#>  [6] -0.062366283 -0.034077531 -0.065866838 -0.077004900 -0.069408707
#> [11] -0.039634368 -0.064516605 -0.040309404 -0.055485515 -0.041468916
#> [16] -0.065604814 -0.048062279 -0.048401972 -0.040263312 -0.037806611
#> [21] -0.036782582 -0.054367900 -0.043462585 -0.049426781 -0.051941112
#> [26] -0.034187142 -0.064187815 -0.043712593 -0.066814169 -0.048289449
#> [31] -0.033417222 -0.051565914 -0.073016649 -0.043667173 -0.023507488
#> [36] -0.052379448 -0.064473907 -0.049087303 -0.025581130 -0.049896364
#> [41] -0.050040422 -0.044987176 -0.048382190 -0.018670728 -0.025563795
#> [46] -0.035773687  0.028568800 -0.020199918 -0.009257894 -0.068670943
#> 
#> $latitude
#>  [1] 51.52158 51.51341 51.50523 51.51306 51.49500 51.49451 51.49612 51.50118
#>  [9] 51.49851 51.50317 51.51825 51.49583 51.51405 51.53783 51.52189 51.49068
#> [17] 51.51682 51.50008 51.52168 51.51430 51.51769 51.52651 51.51365 51.51376
#> [25] 51.49753 51.51094 51.51096 51.52570 51.51757 51.50034 51.50480 51.51000
#> [33] 51.50998 51.51250 51.49303 51.49706 51.50417 51.50625 51.49364 51.49814
#> [41] 51.50318 51.51571 51.48524 51.49301 51.51386 51.50631 51.52748 51.52387
#> [49] 51.51219 51.52918
#>