Create sources data object in same format as observations

Simple function that ensures that sources are in the correct format required by Rgeoprofile. Takes longitude and latitude as input vectors and returns these same values in list format. If no values are input then default values are used.

geoDataSource(longitude = NULL, latitude = NULL)

Arguments

longitude
the locations of the potential sources in degrees longitude.

latitude
the locations of the potential sources in degrees latitude.

Examples

# John Snow cholera data
geoDataSource(WaterPumps$longitude, WaterPumps$latitude)

#> $longitude
#>  [1] -0.139678 -0.139583 -0.136670 -0.139656 -0.138266 -0.135990 -0.133641
#>  [8] -0.131694 -0.139649 -0.137783 -0.135909 -0.134724 -0.134102
#> 
#> $latitude
#>  [1] 51.51486 51.51382 51.51327 51.51007 51.51119 51.51142 51.51202 51.51223
#>  [9] 51.51634 51.51662 51.51605 51.51624 51.50995
#> 

# simulated data
sim <-rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude =
51.5235505, alpha=1, sigma=1, tau=3)
geoDataSource(sim$longitude, sim$latitude)

#> $longitude
#>  [1] -0.029846510 -0.022355893 -0.046943191 -0.018294959 -0.026393270
#>  [6] -0.021512941 -0.054118122 -0.031142845 -0.037608642 -0.112580396
#> [11] -0.111855147 -0.112512923 -0.110434773 -0.118377447 -0.110007153
#> [16] -0.107715154 -0.119158450 -0.118403872 -0.137925202 -0.114840540
#> [21] -0.132466974 -0.111015338 -0.127540469 -0.134437704 -0.029566663
#> [26] -0.049939706 -0.055667819 -0.052709048 -0.025781578 -0.028267914
#> [31] -0.028683632 -0.036947733 -0.020282562 -0.042947639 -0.029982339
#> [36] -0.039208378 -0.040212295 -0.050733174 -0.017120110 -0.040075235
#> [41] -0.003908093 -0.009659131 -0.003329104  0.010787230 -0.004938108
#> [46] -0.009859465 -0.006497051  0.006335018 -0.026988909 -0.003374440
#> 
#> $latitude
#>  [1] 51.49775 51.48022 51.48738 51.49366 51.49222 51.49307 51.48881 51.50331
#>  [9] 51.50147 51.52157 51.51743 51.52424 51.51693 51.52809 51.51109 51.53488
#> [17] 51.51420 51.51474 51.54498 51.53375 51.53047 51.52631 51.51691 51.54206
#> [25] 51.50735 51.50265 51.51595 51.51806 51.50182 51.52213 51.51595 51.50431
#> [33] 51.50355 51.50401 51.50084 51.49724 51.50805 51.51081 51.52780 51.51605
#> [41] 51.50781 51.50504 51.52052 51.51127 51.53775 51.53712 51.54742 51.53402
#> [49] 51.52802 51.53090
#>