Draw from Dirichlet process mixture model

Provides random draws from a 2D spatial Dirichlet process mixture model. Both sigma and tau are defined in units of kilometres, representing the average distance that an observations lies from a source, and the average distance that a source lies from the centre point respectively. In contrast, the location of the centre point and the locations of the final output crime sites are defined in units of degrees lat/long to facilitate spatial analysis.

rDPM(n, sigma = 1, tau = 10, priorMean_longitude = -0.1277,
  priorMean_latitude = 51.5074, alpha = 1)

Arguments

n
number of draws.

sigma
standard deviation of dispersal distribution, in units of kilometres.

tau
standard deviation of prior on source locations (i.e. average distances of sources from centre point), in units of kilometres.

priorMean_longitude
location of prior mean on source locations in degrees longitude.

priorMean_latitude
location of prior mean on source locations in degrees latitude.

alpha
concentration parameter of Dirichlet process model. Large alpha implies many distinct sources, while small alpha implies only a few sources.

Details

Output includes the lat/long locations of the points drawn from the DPM model, along with the underlying group allocation (i.e. which points belong to which sources) and the lat/long locations of the sources.

Examples

# produces clusters of points from sources centred on QMUL
rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 51.5235505,
alpha=1, sigma=1, tau=3)

#> $longitude
#>  [1] -0.0680037418 -0.0572632560 -0.0546916181 -0.0496948482 -0.0603683957
#>  [6] -0.0435778472 -0.0336578667 -0.0127705020 -0.0542437468 -0.0464171569
#> [11] -0.0571265549 -0.0448368650 -0.0282470727 -0.0405508114 -0.0417613380
#> [16] -0.0411051870 -0.0527423279 -0.0424753925 -0.0452879356 -0.0402510291
#> [21] -0.0654839665 -0.0269696759 -0.0654207004 -0.0540948143 -0.0404951604
#> [26] -0.0466309864 -0.0335235143 -0.0465195051 -0.0449914185 -0.0536224065
#> [31] -0.0492545123 -0.0346598721 -0.0490927039 -0.0586789802 -0.0749343509
#> [36] -0.0622261556 -0.0574935031 -0.0350510714 -0.0527861714 -0.0613024611
#> [41]  0.0001819212 -0.0347494734 -0.0217906689 -0.0104324005 -0.0355138328
#> [46]  0.0053698325 -0.0091261848 -0.0107411961 -0.0260607773 -0.0269064271
#> 
#> $latitude
#>  [1] 51.49313 51.48880 51.49832 51.50490 51.50301 51.50482 51.50213 51.51141
#>  [9] 51.50654 51.49930 51.48337 51.49098 51.50659 51.48902 51.50208 51.48725
#> [17] 51.50557 51.49775 51.49276 51.50882 51.50275 51.49242 51.51143 51.50641
#> [25] 51.49818 51.49451 51.50621 51.51205 51.50394 51.51076 51.49671 51.50344
#> [33] 51.52346 51.50306 51.50213 51.50740 51.51529 51.51192 51.50324 51.50798
#> [41] 51.52482 51.52682 51.51737 51.51991 51.54157 51.52765 51.53836 51.54819
#> [49] 51.53239 51.52631
#> 
#> $group
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 2 2 2 2 2 2 2 2 2 2
#> 
#> $source_lon
#> [1] -0.045381011 -0.007500685
#> 
#> $source_lat
#> [1] 51.50276 51.53017
#> 

# same, but increasing alpha to generate more clusters
rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 51.5235505,
alpha=5, sigma=1, tau=3)

#> $longitude
#>  [1] -0.076037063 -0.061775408 -0.052784030 -0.056049798 -0.062883841
#>  [6] -0.038101093 -0.051677759 -0.045833946 -0.055853840 -0.041116226
#> [11] -0.054417792 -0.067498798 -0.069492139 -0.079996778 -0.069581427
#> [16] -0.070502239 -0.048584368 -0.035823222 -0.077956868 -0.055964162
#> [21] -0.068212857 -0.038745857 -0.020717872 -0.060599593 -0.034285576
#> [26] -0.029041504 -0.016953441 -0.036583598 -0.039022744 -0.042627059
#> [31] -0.023957785  0.024585407  0.012743422 -0.005322346  0.001887381
#> [36] -0.002298251 -0.003377656  0.007096733 -0.018380533 -0.058700766
#> [41] -0.057345609 -0.086976600 -0.059496431  0.071497534 -0.134320167
#> [46] -0.022886198 -0.014061554 -0.021942397 -0.017974062 -0.011377129
#> 
#> $latitude
#>  [1] 51.56853 51.55795 51.56637 51.57119 51.58439 51.49149 51.48547 51.47443
#>  [9] 51.49203 51.48160 51.49397 51.48816 51.49670 51.49934 51.48575 51.50518
#> [17] 51.49294 51.50993 51.49592 51.50770 51.49272 51.48493 51.49777 51.56782
#> [25] 51.57048 51.55891 51.58053 51.55873 51.57119 51.57230 51.54457 51.53758
#> [33] 51.55859 51.53006 51.54369 51.53883 51.54455 51.54287 51.55862 51.47685
#> [41] 51.50675 51.49242 51.49864 51.49870 51.49895 51.50573 51.52429 51.52405
#> [49] 51.50120 51.51747
#> 
#> $group
#>  [1]  1  1  1  2  2  3  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  4  5  5
#> [26]  5  5  5  5  5  6  6  6  6  6  6  6  6  7  8  9  9  9 10 11 12 12 12 12 12
#> 
#> $source_lon
#>  [1] -0.063208667 -0.070525794 -0.054908096 -0.054178202 -0.036251555
#>  [6]  0.004116384 -0.042095571 -0.036107153 -0.064737435  0.060363342
#> [11] -0.115699137 -0.012581674
#> 
#> $source_lat
#>  [1] 51.56028 51.57497 51.50540 51.49487 51.56763 51.53813 51.56286 51.47138
#>  [9] 51.49980 51.50198 51.49962 51.51559
#>