Run main MCMC

Run the main RgeoProfile MCMC. Model parameters are taken from the current active parameter set, and MCMC parameters are passed in as arguments. All output is stored within the project.

run_mcmc(project, K = 3, precision_lon = 0.001, precision_lat = 0.001,
  burnin = 100, samples = 1000, auto_converge = TRUE,
  converge_test = ceiling(burnin/10), solve_label_switching_on = TRUE,
  cluster = NULL, pb_markdown = FALSE, store_raw = TRUE,
  silent = !is.null(cluster))

Arguments

project
an rgeoprofile_project, as produced by the function rgeoprofile_project()

K
the number of sources

precision_lon
the level of precision at which longitudes are represented

precision_lat
the level of precision at which latitudes are represented

burnin
the number of burn-in iterations

samples
the number of sampling iterations

auto_converge
whether convergence should be assessed automatically every converge_test iterations, leading to termination of the burn-in phase. If FALSE then the full burnin iterations are used

converge_test
test for convergence every convergence_test iterations if auto_converge is being used

solve_label_switching_on
whether to implement the Stevens' solution to the label-switching problem. If turned off then Q-matrix output will no longer be correct, although evidence estimates will be unaffected

cluster
option to pass in a cluster environment (see package "parallel")

pb_markdown
whether to run progress bars in markdown mode, in which case they are updated once at the end to avoid large amounts of output

store_raw
whether to store raw MCMC output in addition to summary output. Setting to FALSE can considerably reduce output size in memory

silent
whether to suppress all console output

Details

Both longitude and latitude values can be represented to a given precision level using the arguments precision_lon and precision_lat - for example, a precision of 0.01 means that values are rounded to the nearest hundredth of a degree. This allows the use of look-up tables for the likelihood calculation, which significantly speeds up the MCMC. Set to 0 to use exact values (up to C++ "double" precision) rather than using look-up tables.