Draw from Dirichlet process mixture model

Provides random draws from a 2D spatial Dirichlet process mixture model. Both sigma and tau are defined in units of kilometres, representing the average distance that an observations lies from a source, and the average distance that a source lies from the centre point respectively. In contrast, the location of the centre point and the locations of the final output crime sites are defined in units of degrees lat/long to facilitate spatial analysis.

rDPM(n, sigma = 1, tau = 10, priorMean_longitude = -0.1277,
  priorMean_latitude = 51.5074, alpha = 1)

Arguments

n
number of draws.

sigma
standard deviation of dispersal distribution, in units of kilometres.

tau
standard deviation of prior on source locations (i.e. average distances of sources from centre point), in units of kilometres.

priorMean_longitude
location of prior mean on source locations in degrees longitude.

priorMean_latitude
location of prior mean on source locations in degrees latitude.

alpha
concentration parameter of Dirichlet process model. Large alpha implies many distinct sources, while small alpha implies only a few sources.

Details

Output includes the lat/long locations of the points drawn from the DPM model, along with the underlying group allocation (i.e. which points belong to which sources) and the lat/long locations of the sources.

Examples

# produces clusters of points from sources centred on QMUL
rDPM(n = 50, 
     priorMean_longitude = -0.04217491, 
     priorMean_latitude = 51.5235505, 
     alpha = 1, 
     sigma = 1, 
     tau=3) 

#> $longitude
#>  [1]  0.015911972  0.020635633  0.030932319  0.007512090  0.000207557
#>  [6]  0.013089127  0.031109343  0.029226291  0.009833875  0.025433730
#> [11]  0.016074981 -0.001807459  0.030774997  0.004532026  0.019305284
#> [16]  0.054426683  0.027640513  0.019550080  0.031517833  0.061611116
#> [21]  0.038540349  0.038487076  0.023795752  0.021483339  0.053596785
#> [26]  0.024696305  0.010859860  0.028877634  0.035058505  0.027765526
#> [31]  0.021377732  0.031699488  0.004110612 -0.010719840  0.032834088
#> [36]  0.019793949  0.025491448  0.017842003  0.029991304  0.028249948
#> [41]  0.029395415  0.008099270  0.016776026 -0.057218502 -0.059329722
#> [46] -0.058161699 -0.055411890 -0.037205594 -0.042899725 -0.037959925
#> 
#> $latitude
#>  [1] 51.52315 51.52247 51.53361 51.51403 51.53376 51.53523 51.53552 51.52751
#>  [9] 51.51995 51.53106 51.54263 51.53655 51.52406 51.51455 51.51267 51.53004
#> [17] 51.53921 51.52650 51.51433 51.53168 51.53247 51.52058 51.53439 51.52300
#> [25] 51.52712 51.52792 51.53150 51.54293 51.52119 51.52627 51.51536 51.52362
#> [33] 51.53972 51.52189 51.52965 51.53698 51.52228 51.53845 51.55058 51.54399
#> [41] 51.51457 51.54108 51.52477 51.55339 51.55545 51.53290 51.55984 51.53485
#> [49] 51.55891 51.56453
#> 
#> $group
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [39] 1 1 1 1 1 2 2 2 2 2 2 2
#> 
#> $source_lon
#> [1]  0.02636628 -0.05079661
#> 
#> $source_lat
#> [1] 51.52834 51.54898
#> 

# same, but increasing alpha to generate more clusters
rDPM(n = 50, 
     priorMean_longitude = -0.04217491, 
     priorMean_latitude = 51.5235505, 
     alpha = 5, 
     sigma = 1, 
     tau = 3)

#> $longitude
#>  [1] -0.030001067 -0.011575682 -0.029344048 -0.003628818 -0.006011755
#>  [6] -0.013083432 -0.001889009 -0.013799373 -0.009061030  0.004800509
#> [11]  0.010663393 -0.020631491  0.009049799  0.013348334 -0.014139959
#> [16] -0.020055154  0.001065718 -0.112943224 -0.090373669 -0.092089386
#> [21] -0.096864171 -0.102049467 -0.112483707 -0.092693605 -0.083734898
#> [26] -0.079813546 -0.095918229 -0.110988427 -0.080022975 -0.071180686
#> [31] -0.086661074 -0.068621297 -0.083204996 -0.084220185 -0.112395967
#> [36] -0.102863614 -0.034401825 -0.057808995 -0.047026118 -0.036500757
#> [41] -0.054502689 -0.062083145 -0.041607205 -0.032517017 -0.045606203
#> [46] -0.017265125 -0.092840012 -0.033905015 -0.110721251 -0.088887371
#> 
#> $latitude
#>  [1] 51.58142 51.58531 51.56819 51.56365 51.58443 51.58617 51.58653 51.57642
#>  [9] 51.58562 51.57360 51.57330 51.57283 51.56186 51.56855 51.57397 51.58028
#> [17] 51.57749 51.48375 51.46884 51.47790 51.48634 51.47287 51.47004 51.47831
#> [25] 51.46643 51.50120 51.49199 51.50434 51.50897 51.49051 51.50529 51.49517
#> [33] 51.48871 51.50109 51.46919 51.45593 51.51919 51.49461 51.47657 51.46899
#> [41] 51.49686 51.50716 51.49278 51.51950 51.51989 51.50128 51.53844 51.53182
#> [49] 51.47928 51.48944
#> 
#> $group
#>  [1]  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  2  2  2  2  2  2  2
#> [26]  3  3  3  3  3  3  3  3  3  4  4  5  5  6  6  7  7  7  8  8  9 10 11 12 13
#> 
#> $source_lon
#>  [1] -0.004214453 -0.098386000 -0.088935998 -0.091327667 -0.041194769
#>  [6] -0.043104486 -0.054257721 -0.043224185 -0.006415945 -0.074717850
#> [11] -0.018924469 -0.110296157 -0.084916290
#> 
#> $source_lat
#>  [1] 51.57706 51.47359 51.49900 51.47089 51.50196 51.47312 51.50330 51.50761
#>  [9] 51.50352 51.54226 51.52194 51.47324 51.48916
#>