Draw from Dirichlet process mixture model

Provides random draws from a 2D spatial Dirichlet process mixture model. Both sigma and tau are defined in units of kilometres, representing the average distance that an observations lies from a source, and the average distance that a source lies from the centre point respectively. In contrast, the location of the centre point and the locations of the final output crime sites are defined in units of degrees lat/long to facilitate spatial analysis.

rDPM(n, sigma = 1, tau = 10, priorMean_longitude = -0.1277,
  priorMean_latitude = 51.5074, alpha = 1)

Arguments

n
number of draws.

sigma
standard deviation of dispersal distribution, in units of kilometres.

tau
standard deviation of prior on source locations (i.e. average distances of sources from centre point), in units of kilometres.

priorMean_longitude
location of prior mean on source locations in degrees longitude.

priorMean_latitude
location of prior mean on source locations in degrees latitude.

alpha
concentration parameter of Dirichlet process model. Large alpha implies many distinct sources, while small alpha implies only a few sources.

Details

Output includes the lat/long locations of the points drawn from the DPM model, along with the underlying group allocation (i.e. which points belong to which sources) and the lat/long locations of the sources.

Examples

# produces clusters of points from sources centred on QMUL
rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 51.5235505, 
alpha=1, sigma=1, tau=3) 

#> $longitude
#>  [1] -0.07071854 -0.02899804 -0.08771382 -0.06063891 -0.05081275 -0.04295398
#>  [7] -0.08024408 -0.08588012 -0.04688169 -0.06268836 -0.04944622 -0.09847647
#> [13] -0.08324485 -0.08204795 -0.08332704 -0.07477592 -0.08333785 -0.06907868
#> [19] -0.06008887 -0.05604290 -0.04727986 -0.03809947 -0.05241117 -0.04656367
#> [25] -0.05658988 -0.04184298 -0.05515293 -0.06824218 -0.07023678 -0.08074805
#> [31] -0.07032612 -0.07124752 -0.04931582 -0.03654662 -0.07870685 -0.05670027
#> [37] -0.06895669 -0.03947111 -0.02143175 -0.07923767 -0.05294911 -0.04771011
#> [43] -0.03563374 -0.05524491 -0.05768169 -0.06055662 -0.08225426 -0.03372259
#> [49] -0.04556177 -0.06362325
#> 
#> $latitude
#>  [1] 51.52584 51.52324 51.52257 51.55419 51.56708 51.53826 51.55694 51.53849
#>  [9] 51.55560 51.55927 51.54075 51.54592 51.55142 51.54652 51.54898 51.55309
#> [17] 51.56059 51.55001 51.55843 51.54856 51.53302 51.50969 51.51419 51.50316
#> [25] 51.52075 51.51032 51.52270 51.51689 51.52543 51.52807 51.51448 51.53390
#> [33] 51.52167 51.53866 51.52464 51.53643 51.52144 51.51366 51.52650 51.52379
#> [41] 51.52646 51.51488 51.53650 51.51471 51.52716 51.53201 51.53377 51.52678
#> [49] 51.54778 51.51926
#> 
#> $group
#>  [1] 1 1 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3
#> [39] 3 3 3 3 3 3 3 4 4 4 4 4
#> 
#> $source_lon
#> [1] -0.06319340 -0.07051169 -0.05491319 -0.05418676
#> 
#> $source_lat
#> [1] 51.52723 51.55234 51.52360 51.52733
#> 

# same, but increasing alpha to generate more clusters
rDPM(50, priorMean_longitude = -0.04217491, priorMean_latitude = 51.5235505, 
alpha=5, sigma=1, tau=3)

#> $longitude
#>  [1] -0.061458998 -0.065096832 -0.090627180 -0.033639377 -0.067317680
#>  [6] -0.072873141 -0.054903906 -0.088553324 -0.052818062 -0.068406517
#> [11] -0.068670395 -0.075435216 -0.077219666 -0.093906576 -0.077233993
#> [16] -0.057958761 -0.063142057 -0.077650794 -0.056754804 -0.084389700
#> [21] -0.181289807 -0.043056277 -0.051620131 -0.058691234 -0.059084960
#> [26] -0.054965546 -0.049463734 -0.092965203 -0.072625624 -0.028639055
#> [31] -0.042124330 -0.065138739 -0.098644358 -0.008862761  0.001735281
#> [36] -0.058866670 -0.054428014 -0.062696864  0.083018383  0.052715845
#> [41]  0.044991475  0.051404010  0.053632711 -0.014155753 -0.004476909
#> [46] -0.017540070 -0.038706409 -0.067930741 -0.046534400 -0.064108698
#> 
#> $latitude
#>  [1] 51.49486 51.48571 51.49031 51.48818 51.48582 51.49876 51.49946 51.50339
#>  [9] 51.48312 51.48938 51.51359 51.49072 51.49058 51.49709 51.49612 51.49460
#> [17] 51.49490 51.49851 51.49414 51.50446 51.51423 51.52036 51.53052 51.53469
#> [25] 51.54347 51.53796 51.53846 51.52933 51.55056 51.53481 51.53678 51.53571
#> [33] 51.53803 51.54142 51.53741 51.46538 51.46673 51.46022 51.50970 51.51586
#> [41] 51.51418 51.51411 51.49698 51.47799 51.55014 51.53707 51.54577 51.54968
#> [49] 51.54694 51.55003
#> 
#> $group
#>  [1]  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  3  3  3  3
#> [26]  3  3  3  3  3  3  3  4  5  5  6  6  6  7  7  7  7  7  8  9  9 10 10 10 11
#> 
#> $source_lon
#>  [1] -0.0692210482 -0.1613950217 -0.0499750105 -0.0897163773 -0.0008425913
#>  [6] -0.0708127995  0.0571507630 -0.0240175762 -0.0169510090 -0.0511232717
#> [11] -0.0465658585
#> 
#> $source_lat
#>  [1] 51.49346 51.50904 51.53707 51.52453 51.53592 51.48248 51.51158 51.47571
#>  [9] 51.53809 51.54304 51.53928
#>